+++ src/main/java/ru/practicum/shareit/booking/repository/BookingRepository.java
@@ -10,52 +10,52 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
-    List<Booking> findAllByBookerId(Integer bookerId);
+public interface BookingRepository extends JpaRepository<Booking, Long> {
+    List<Booking> findAllByBookerId(Long bookerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.booker.id = :bookerId " +
             "AND current_timestamp < b.start")
-    List<Booking> findByBookerAndStateCurrent(Integer bookerId);
+    List<Booking> findByBookerAndStateCurrent(Long bookerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.booker.id = :bookerId " +
             "AND current_timestamp > b.end")
-    List<Booking> findByBookerAndStatePast(Integer bookerId);
+    List<Booking> findByBookerAndStatePast(Long bookerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.booker.id = :bookerId " +
             "AND current_timestamp < b.start")
-    List<Booking> findByBookerAndStateFuture(Integer bookerId);
+    List<Booking> findByBookerAndStateFuture(Long bookerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.booker.id = :bookerId " +
             "AND b.status = :status")
-    List<Booking> findAllByBookerIdAndStatus(Integer bookerId, Status status);
+    List<Booking> findAllByBookerIdAndStatus(Long bookerId, Status status);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.item.owner.id = :ownerId")
-    List<Booking> findAllByOwnerId(Integer ownerId);
+    List<Booking> findAllByOwnerId(Long ownerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.item.owner.id = :ownerId " +
             "AND current_timestamp < b.start")
-    List<Booking> findByOwnerAndStateCurrent(Integer ownerId);
+    List<Booking> findByOwnerAndStateCurrent(Long ownerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.item.owner.id = :ownerId " +
             "AND current_timestamp > b.end")
-    List<Booking> findByOwnerAndStatePast(Integer ownerId);
+    List<Booking> findByOwnerAndStatePast(Long ownerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.item.owner.id = :ownerId " +
             "AND current_timestamp < b.start")
-    List<Booking> findByOwnerAndStateFuture(Integer ownerId);
+    List<Booking> findByOwnerAndStateFuture(Long ownerId);
 
     @Query("SELECT b FROM Booking b " +
             "WHERE b.item.owner.id = :ownerId " +
             "AND b.status = :status")
-    List<Booking> findAllByOwnerIdAndStatus(Integer ownerId, Status status);
+    List<Booking> findAllByOwnerIdAndStatus(Long ownerId, Status status);
 
-    List<Booking> findAllByBookerIdAndItemIdAndStatusEqualsAndEndIsBefore(Integer userId, Integer itemId, Status status, LocalDateTime now);
+    List<Booking> findAllByBookerIdAndItemIdAndStatusEqualsAndEndIsBefore(Long userId, Long itemId, Status status, LocalDateTime now);
 }
