+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -6,7 +6,6 @@ import lombok.*;
 import ru.practicum.shareit.request.model.ItemRequest;
 import ru.practicum.shareit.user.model.User;
 
-
 @Getter
 @Setter
 @Entity
@@ -14,12 +13,12 @@ import ru.practicum.shareit.user.model.User;
 @Table(name = "items")
 @NoArgsConstructor
 @AllArgsConstructor
-@JsonIgnoreProperties(value = { "hibernateLazyInitializer", "handler" })
+@JsonIgnoreProperties(value = {"hibernateLazyInitializer", "handler"})
 public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name", nullable = false, length = 255)
     private String name;
     @Column(name = "description", nullable = false, length = 1024)
